Create general purpose confirmation dialog with password input
